+++ test/utils.js
@@ -102,6 +102,8 @@ window.test = {
 
   'LiveChat' : '2328351',
 
+  'Lytics' : 'x',
+
   'Mixpanel' : 'x',
 
   'Olark' : '5798-949-10-1692',